+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * Rename Mock to Capture
 * Add `(*=>)` combinator
 * Add heartbeat usage example
+* Add `bypass` combinator